requirements copy.txt
python-telegram-bot>=20.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
mongoengine>=0.24.0
pymongo>=4.0.0
dnspython>=2.2.0
flask>=2.0.0
flask-cors>=3.0.10